Consolidate fft1d, fft3d, rsgs classes into single BasisVectorSearch class inheriting from indexer_base
